+++ src/TryCourier.Tests/Models/WorkspacePreferences/WorkspacePreferenceCreateParamsTest.cs
@@ -14,11 +14,13 @@ public class WorkspacePreferenceCreateParamsTest : TestBase
         var parameters = new WorkspacePreferenceCreateParams
         {
             Name = "Account Notifications",
+            Description = "description",
             HasCustomRouting = true,
             RoutingOptions = [ChannelClassification.DirectMessage],
         };
 
         string expectedName = "Account Notifications";
+        string expectedDescription = "description";
         bool expectedHasCustomRouting = true;
         List<ApiEnum<string, ChannelClassification>> expectedRoutingOptions =
         [
@@ -26,6 +28,7 @@ public class WorkspacePreferenceCreateParamsTest : TestBase
         ];
 
         Assert.Equal(expectedName, parameters.Name);
+        Assert.Equal(expectedDescription, parameters.Description);
         Assert.Equal(expectedHasCustomRouting, parameters.HasCustomRouting);
         Assert.NotNull(parameters.RoutingOptions);
         Assert.Equal(expectedRoutingOptions.Count, parameters.RoutingOptions.Count);
@@ -40,6 +43,8 @@ public class WorkspacePreferenceCreateParamsTest : TestBase
     {
         var parameters = new WorkspacePreferenceCreateParams { Name = "Account Notifications" };
 
+        Assert.Null(parameters.Description);
+        Assert.False(parameters.RawBodyData.ContainsKey("description"));
         Assert.Null(parameters.HasCustomRouting);
         Assert.False(parameters.RawBodyData.ContainsKey("has_custom_routing"));
         Assert.Null(parameters.RoutingOptions);
@@ -53,10 +58,13 @@ public class WorkspacePreferenceCreateParamsTest : TestBase
         {
             Name = "Account Notifications",
 
+            Description = null,
             HasCustomRouting = null,
             RoutingOptions = null,
         };
 
+        Assert.Null(parameters.Description);
+        Assert.True(parameters.RawBodyData.ContainsKey("description"));
         Assert.Null(parameters.HasCustomRouting);
         Assert.True(parameters.RawBodyData.ContainsKey("has_custom_routing"));
         Assert.Null(parameters.RoutingOptions);
@@ -81,6 +89,7 @@ public class WorkspacePreferenceCreateParamsTest : TestBase
         var parameters = new WorkspacePreferenceCreateParams
         {
             Name = "Account Notifications",
+            Description = "description",
             HasCustomRouting = true,
             RoutingOptions = [ChannelClassification.DirectMessage],
         };